feature(sessionprojector): add AVFoundation-based ScreenRecorder implementation (which works on loginscreen)
